Reworked classes.challenge (and challenges/ dir)
Challenges dir is now under classes dir

Changed how challenges were coded and imported
Now they can be called as such: Challenge.basic(), Challenge.ngu(), etc.
